backend/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
pyspark==3.5.0
delta-spark==3.0.0
pyiceberg==0.5.1
boto3==1.34.0
python-dotenv==1.0.0
